requirements.txt
Django==1.8.6
numpy==1.10.1
Pillow==3.0.0
qrcode==5.1
reportlab==3.2.0
six==1.10.0
wheel==0.24.0
zbar==0.10
